requirements.txt
langgraph
langchain
openai

wikipedia
duckduckgo-search
ddgs
tavily-python
requests

sentence-transformers

faster-whisper
sounddevice
numpy==1.23.5
spacy<3.7.0
wordfreq
google-generativeai

python-dotenv
rich
